only clean up if _testing file exists

+++ tests/test_paper.py
@@ -36,8 +36,9 @@ log = get_logger(__name__)
 def paper_cleanup():
     yield
     app_dir = get_app_dir('piker')
-    rmtree(app_dir)
-    assert not os.path.isfile(app_dir)
+    if os.path.isfile(app_dir)        
+        rmtree(app_dir)
+        assert not os.path.isfile(app_dir)
 
 def test_paper_trade(
     open_test_pikerd: AsyncContextManager,
